refactor: service 패키지를 application패키지로 이름 변경
